requirements.txt
tornado
google-generativeai
python-dotenv
matplotlib
matplotlib-venn
